feat: 로컬 개발용 기능을 간소화
