--- backend/src/main/resources/local/import.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8"/>
-    <title>파일 업로드</title>
-</head>
-<body>
-<h1>파일 업로드 폼</h1>
-<form action="/admin/courses/import" method="post" enctype="multipart/form-data">
-    <label for="files">파싱할 파일 전송:</label>
-    <input type="file" id="files" name="files" multiple required />
-
-    <button type="submit">전송</button>
-</form>
-</body>
-</html>
